# improve lock perf