== before change:
BenchmarkLock-8   	 2000000	       967 ns/op
BenchmarkLock-8   	 2000000	       966 ns/op
BenchmarkLock-8   	 2000000	       983 ns/op
BenchmarkLock-8   	 2000000	       969 ns/op

==  after change:
BenchmarkLock-8   	 2000000	       669 ns/op
BenchmarkLock-8   	 2000000	       679 ns/op
BenchmarkLock-8   	 2000000	       690 ns/op
BenchmarkLock-8   	 2000000	       671 ns/op